staging progress
updated framerate, probably need to make it variable or pause animations or bypass animations for effects that don't need it

breathing isnt' correct, but the counting effect is neat, may reuse it for something else

the game class is  non functional, needs to be turned into a simple state machine that looks for different pattern combos

keeping the current menu state machine since it allows for 11 different mini pattern games

todo:
need to make the reset more responsive, I tried to implement a hold function to reset but that didn't work as intended
maybe just reduce the click duration to 1 second?

tweak the framerate, it's not correct and takes too long to cycle through

come up with 10 unique blink patterns or led effects that are interesting for each unused state.